etcdctl: fix lint and build errors in diagnosis feature
- Fix markdown lint errors (MD024, MD004, MD007, MD012, MD022) in DIAGNOSIS section only
- Fix GCI import formatting in diagnosis plugins
- Fix 32-bit overflow by changing dbQuotaBytes from int to int64
- Remove unnecessary int64() conversion in epstatus plugin

+++ etcdctl/ctlv3/command/diagnosis/plugins/common/checker.go
@@ -27,6 +27,6 @@ type Checker struct {
 	Cfg            *clientv3.ConfigSpec
 	Endpoints      []string
 	CommandTimeout time.Duration
-	DbQuotaBytes   int
+	DbQuotaBytes   int64
 	Name           string
 }
